Conflicts:
	quadcopter/src/drive.cpp
	quadcopter/tests/visualize/Makefile
	quadcopter/tests/visualize/test_visualizedrive.cpp

+++ quadcopter/tests/test_gyroscope.cpp
@@ -22,7 +22,7 @@ int main(int argc, char **argv) {
 
 		int numsamples = 5;
 		int i = 1;
-		Vector3<int> values;
+		Vector3<float> values;
 
 		std::string input;
 		bool running = true;
